+++ lib/services/location_service.dart
@@ -33,7 +33,8 @@ class LocationService {
 
       return (latitude: position.latitude, longitude: position.longitude);
     } catch (e, stack) {
-      developer.log('Ошибка при получении геолокации', error: e, stackTrace: stack, name: 'LocationService');
+      developer.log('Ошибка при получении геолокации',
+          error: e, stackTrace: stack, name: 'LocationService');
       return null;
     }
   }